requirements.txt
aniso8601==7.0.0
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
dateparser==0.7.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.23.0
flask-restplus==0.13.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
mongoengine==0.18.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.9.0
pyOpenSSL==19.0.0
pyrsistent==0.15.4
python-binance==0.7.3
python-dateutil==2.8.0
pytz==2019.2
regex==2019.8.19
requests==2.22.0
service-identity==18.1.0
six==1.12.0
Twisted==19.7.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.6
zope.interface==4.6.0
